refactor: jwt movido para sua propria função
