--- src/auth/authService.js
@@ -1,26 +1,0 @@
-const jwt = require('jsonwebtoken');
-const bcrypt = require('bcrypt');
-const prisma = require('../prisma');
-
-async function loginUser(email, password) {
-  const user = await prisma.user.findUnique({
-    where: { email },
-  });
-
-  if (!user) {
-    throw new Error('Usuário não encontrado.');
-  }
-
-  const isPasswordValid = await bcrypt.compare(password, user.password);
-  if (!isPasswordValid) {
-    throw new Error('Email ou senha incorretos.');
-  }
-
-  const token = jwt.sign({ id: user.id, email: user.email }, process.env.JWT_SECRET, {
-    expiresIn: '1h',
-  });
-
-  return { user: { id: user.id, name: user.name, email: user.email }, token };
-}
-
-module.exports = { loginUser };
